requirements.txt
PyMuPDF 
Pillow
nltk